requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
scikit-learn
xgboost
lime
shap
beautifulsoup4
nltk

wordsegment
wordcloud
contractions
pyspellchecker
scipy
ipython
pickleshare
wordninja

# spacy==3.5.0
# thinc==8.1.10

# # SpaCy Model
# https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
#en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0.tar.gz
